+++ R/02-get_regional_column.R
@@ -5,11 +5,11 @@
 #' @noRd
 
 .get_regional_column <- function(geography) {
-  stopifnot(
-    "`geography` must be either `country`, `region`, `income` or `world`" =
-      is.character(match.arg(geography, choices =
-                               .pkgenv$wrp$wrp_regions$WRP_UID))
-  )
+  if (!(geography %in% .pkgenv$wrp$wrp_regions$WRP_UID)) {
+    message("INPUT ERROR: `geography` must be either `country`, `region`, 
+            `income` or `world`.\nSee `?wrp_get` for help.")
+    return(-1)
+  }
   geography <- .pkgenv$wrp$wrp_regions$pos[match(geography,
                                                  .pkgenv$wrp$wrp_regions$WRP_UID)]
   return(geography)