+++ R/06-wrp_clean.R
@@ -6,6 +6,7 @@
 #'
 #' @importFrom dplyr mutate mutate_if
 #' @importFrom janitor clean_names
+#' @importFrom labelled remove_attributes
 #'
 #' @return data frame with aggregated World Risk Poll question data
 #' @noRd
@@ -19,6 +20,8 @@ wrp_clean <- function(df) {
     mutate_if(is.factor, as.character) %>%
     clean_names("lower_camel")
   names(df) <- gsub("yearOfInterview", "year", names(df))
-  df <- df[, c(1, 6, 2, 3, 7, 4, 5)]
+  df$year <- as.numeric(df$year)
+  df <- df[, c(1, 5, 2, 3, 7, 4, 6)]
+  df <- df %>% remove_attributes("label")
   return(df)
 }