+++ R/zzz.R
@@ -6,7 +6,7 @@
   test <- has_internet()
   stopifnot("\nNo internet connection detected.\n...You need an active internet connection to use `worldriskpollr`.\n...Once connected, reload the package with `library(worldriskpollr)`" =
               test)
-  url <- "https://github.com/githubIEP/worldriskpollr/raw/master/data-raw/sysdata.rda"
+  url <- "https://github.com/githubIEP/worldriskpollr/raw/master/data-raw/sysdata073.rda"
   url_exists <- HEAD(url)
   stopifnot("\nWorld Risk Poll file not found.\n...Please contact the package maintainer." =
               (url_exists$status_code == 200))