+++ data-raw/format_wrp_from_raw.R
@@ -30,11 +30,12 @@ mutate_if(haven::is.labelled, haven::as_factor)
 wrp_data$countryIncomeLevel <- as.character(wrp_data$countryIncomeLevel2019)
 wrp_data$countryIncomeLevel <- ifelse(is.na(wrp_data$countryIncomeLevel),
                                       as.character(wrp_data$countryIncomeLevel2021), wrp_data$countryIncomeLevel)
+set_label(wrp_data$countryIncomeLevel) <- "countryIncomeLevel"
 wrp_data$world <- "World"
-wrp_data$global <- "Global Statistic"
+set_label(wrp_data$world) <- "World"
 wrp_dictionary <- labelled::generate_dictionary(wrp_data) %>%
   mutate(regional_disaggregate = pos %in% c(2, 6, 233, 234)) %>%
-  mutate(disaggregator = pos %in% c(3, 13:21, 235)) %>%
+  mutate(disaggregator = pos %in% c(13:21)) %>%
   mutate(question = substr(variable, 1, 1) == "q" |
            substr(variable, 1, 2) == "vh") %>%
   mutate(needed = variable %in% c("year", "wgt", "projectionWeight"))
@@ -44,8 +45,8 @@ wrp_dictionary$label <- ifelse((wrp_dictionary$variable == "projectionWeight"),
                                "projectionWeight", wrp_dictionary$label)
 wrp_dictionary$label <- ifelse((wrp_dictionary$variable == "world"),
                                "World", wrp_dictionary$label)
-wrp_dictionary$label <- ifelse((wrp_dictionary$variable == "global"),
-                               "Global Statistic", wrp_dictionary$label)
+# wrp_dictionary$label <- ifelse((wrp_dictionary$variable == "global"),
+#                                "Global Statistic", wrp_dictionary$label)
 wrp_dictionary$label <- ifelse((wrp_dictionary$variable == "projectionWeight"),
                                "projectionWeight", wrp_dictionary$label)
 wrp_dictionary <- wrp_dictionary[wrp_dictionary$regional_disaggregate |
@@ -85,4 +86,4 @@ wrp <- list(
   "wrp_disaggregations" = wrp_disaggregations,
   "wrp_questions" = wrp_questions
 )
-saveRDS(wrp, file = "./data-raw/sysdata.rda", compress = "xz")
+saveRDS(wrp, file = "./data-raw/sysdata073.rda", compress = "xz")